Refactoring to move functions having to do with the scheduler into the scheduler.rs file.
